backend/requirements.txt
# Web Framework
fastapi
uvicorn
gunicorn

# Google Cloud & AI
google-cloud-aiplatform
google-generativeai
firebase-admin
google-cloud-storage
google-auth

# Document Processing
pymupdf4llm
python-docx
mammoth
PyMuPDF
